extract sparkline code to another component
